BE-735 Change to run Explorer from souce tree, not docker in e2e-test
When using docker to start Explorer in Azure pipeline, the result is a little bit flaky
