+++ app/platform/fabric/e2e-test/configs/config_multi.json
@@ -2,11 +2,11 @@
 	"network-configs": {
 		"org1-network": {
 			"name": "org1-network",
-			"profile": "./connection-profile/org1-network.json"
+			"profile": "./e2e-test/configs/connection-profile/org1-network.json"
 		},
 		"org2-network": {
 			"name": "org2-network",
-			"profile": "./connection-profile/org2-network.json"
+			"profile": "./e2e-test/configs/connection-profile/org2-network.json"
 		}
 	},
 	"license": "Apache-2.0"
