+++ app/platform/fabric/e2e-test/configs/config_single.json
@@ -2,7 +2,7 @@
 	"network-configs": {
 		"org1-network": {
 			"name": "org1-network",
-			"profile": "./connection-profile/org1-network.json"
+			"profile": "./e2e-test/configs/connection-profile/org1-network.json"
 		}
 	},
 	"license": "Apache-2.0"
