+++ app/platform/fabric/e2e-test/docker-compose.yaml
@@ -30,6 +30,8 @@ services:
       retries: 5
     volumes:
       - pgdata:/var/lib/postgresql/data
+    ports:
+      - 5432:5432
     networks:
       - mynetwork.com
