+++ app/platform/fabric/e2e-test/specs/apitest_suite_test.go
@@ -1,10 +1,8 @@
 package apitest
 
 import (
-	"strings"
 	"testing"
 
-	"github.com/hyperledger/fabric-test/tools/operator/networkclient"
 	. "github.com/onsi/ginkgo"
 	"github.com/onsi/ginkgo/reporters"
 	. "github.com/onsi/gomega"
@@ -23,21 +21,4 @@ var _ = BeforeSuite(func() {
 // Cleaning up network launched from BeforeSuite and removing all chaincode containers
 // and chaincode container images using AfterSuite
 var _ = AfterSuite(func() {
-
-	dockerList := []string{"ps", "-aq", "-f", "status=exited"}
-	containerList, _ := networkclient.ExecuteCommand("docker", dockerList, false)
-	if containerList != "" {
-		list := strings.Split(containerList, "\n")
-		containerArgs := []string{"rm", "-f"}
-		containerArgs = append(containerArgs, list...)
-		networkclient.ExecuteCommand("docker", containerArgs, true)
-	}
-	ccimagesList := []string{"images", "-q", "--filter=reference=dev*"}
-	images, _ := networkclient.ExecuteCommand("docker", ccimagesList, false)
-	if images != "" {
-		list := strings.Split(images, "\n")
-		imageArgs := []string{"rmi", "-f"}
-		imageArgs = append(imageArgs, list...)
-		networkclient.ExecuteCommand("docker", imageArgs, true)
-	}
 })
