+++ app/platform/fabric/e2e-test/specs/stopexplorer.sh
@@ -3,9 +3,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 SCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
+ROOTPATH="$( cd "$(dirname "$0")/../../../../.." >/dev/null 2>&1 ; pwd -P )"
 
-pushd ${SCRIPTPATH}/..
+pushd ${ROOTPATH}
 
-docker-compose down -v
+./stop.sh
+docker-compose -f ./app/platform/fabric/e2e-test/docker-compose.yaml down -v
 
 popd
